// supabase.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      Friend_request: {
        Row: {
          created_at: string
          id: number
          request_to: string
          sender_id: string
          status: Database["public"]["Enums"]["friendrequeststatusenum"]
        }
        Insert: {
          created_at?: string
          id?: number
          request_to: string
          sender_id: string
          status?: Database["public"]["Enums"]["friendrequeststatusenum"]
        }
        Update: {
          created_at?: string
          id?: number
          request_to?: string
          sender_id?: string
          status?: Database["public"]["Enums"]["friendrequeststatusenum"]
        }
        Relationships: [
          {
            foreignKeyName: "Friend_request_sender_id_fkey"
            columns: ["sender_id"]
            isOneToOne: false
            referencedRelation: "Profile"
            referencedColumns: ["id"]
          },
        ]
      }
      Friendship: {
        Row: {
          created_at: string
          friend_id: string
          id: number
          user_id: string
        }
        Insert: {
          created_at?: string
          friend_id: string
          id?: number
          user_id: string
        }
        Update: {
          created_at?: string
          friend_id?: string
          id?: number
          user_id?: string
        }
        Relationships: []
      }
      Profile: {
        Row: {
          created_at: string
          email: string
          fullname: string | null
          id: string
          updated_at: string
        }
        Insert: {
          created_at?: string
          email: string
          fullname?: string | null
          id: string
          updated_at?: string
        }
        Update: {
          created_at?: string
          email?: string
          fullname?: string | null
          id?: string
          updated_at?: string
        }
        Relationships: []
      }
      Secret_message: {
        Row: {
          author_id: string
          created_at: string
          id: number
          message: string | null
          updated_at: string | null
        }
        Insert: {
          author_id: string
          created_at?: string
          id?: number
          message?: string | null
          updated_at?: string | null
        }
        Update: {
          author_id?: string
          created_at?: string
          id?: number
          message?: string | null
          updated_at?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      friendrequeststatusenum: "PENDING" | "ACCEPTED" | "DECLINED"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never